requirements.txt
langchain-google-genai
chromadb
pandas
python-dotenv
numpy
streamlit
langchain
langchain-community